+++ generators/server/files.js
@@ -13,6 +13,7 @@ const serverFiles = {
             templates: [
                 'src/web/rest/user.controller.ts',
                 'src/web/rest/account.controller.ts',
+                'src/service/dto/password-change.dto.ts',
                 'src/web/rest/auth.controller.ts',
                 'src/web/rest/management.controller.ts',
                 'src/repository/user.repository.ts',