+++ test-integration/01-generate-app-sample.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/sh
 
 set -e
 
@@ -9,7 +9,7 @@ GREEN='\033[0;32m'
 # Change in template directory
 #-------------------------------------------------------------------------------
 cd test-integration/samples/$1
-echo "*** changed directory in : test-integration/samples/"$1
+echo "***${GREEN}changed directory in : test-integration/samples/"$1
 
 #-------------------------------------------------------------------------------
 # Link nodejs blueprint in folder