+++ test-integration/04-run-generated-unit-tests-of-sample.sh
@@ -1,19 +1,19 @@
-#!/bin/bash
+#!/bin/sh
 
 set -e
 
 RED='\033[0;31m'
-
+GREEN='\033[0;32m'
 
 #-------------------------------------------------------------------------------
 # Change in template directory
 #-------------------------------------------------------------------------------
 cd test-integration/samples/$1
-echo "*** changed directory in : test-integration/samples/"$1
+echo "***${GREEN}changed directory in : test-integration/samples/"$1
 
 
 #-------------------------------------------------------------------------------
-# Run unit test 
+# Run unit test
 #-------------------------------------------------------------------------------
 echo "*** run unit test in client for : "$1
 npm run lint:fix && npm test